requirements.txt
loguru
py-cord
flask[async]
flask-discord
hypercorn
python-dotenv
